requirements.txt
Flask==2.3.2
markdown==3.4.3
python-dotenv==1.0.0
openai==0.27.8
requests==2.31.0
